refactor(BusinessHours): extract schedule/rules logic into pure module
Decouples the BusinessHours display component from the editor bundle:
scheduleToRules, rulesToSchedule, and shared types now live in
scheduleRules.ts (no React/component imports). BusinessHoursEditor
re-exports the same names, so the public API is unchanged.
