+++ src/components/BusinessHours/BusinessHours.tsx
@@ -3,7 +3,7 @@
 import * as React from 'react';
 import { cva, type VariantProps } from 'class-variance-authority';
 import { cn } from '../../utils/cn';
-import { scheduleToRules } from '../BusinessHoursEditor/BusinessHoursEditor';
+import { scheduleToRules } from '../BusinessHoursEditor/scheduleRules';
 
 // ============================================================================
 // Types
