+++ src/components/BusinessHoursEditor/BusinessHoursEditor.tsx
@@ -7,22 +7,20 @@ import { Input } from '../Input/Input';
 import { DateInput } from '../DateInput';
 import { Dropdown, DropdownItem } from '../Dropdown';
 import { cn } from '../../utils/cn';
+import {
+  generateId,
+  rulesToSchedule,
+  scheduleToRules,
+  type DaySchedule,
+  type HoursRule,
+} from './scheduleRules';
 
 // ============================================================================
 // Types
 // ============================================================================
 
-export interface TimeSlot {
-  id?: string;
-  start: string;
-  end: string;
-  description?: string;
-}
-
-export interface DaySchedule {
-  day: number; // 0-6 (Sunday to Saturday)
-  hours: TimeSlot[];
-}
+export type { DaySchedule, TimeSlot } from './scheduleRules';
+export { rulesToSchedule, scheduleToRules } from './scheduleRules';
 
 export interface BusinessHoursEditorProps {
   /** Current schedule data */
@@ -69,10 +67,6 @@ const DAY_NAMES_SHORT = ['Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat'];
 // ============================================================================
 // Utilities
 // ============================================================================
-
-function generateId(): string {
-  return Math.random().toString(36).substring(2, 9);
-}
 
 function getOrderedDays(weekStartsOn: 0 | 1): number[] {
   if (weekStartsOn === 1) {
@@ -403,64 +397,6 @@ function BusinessHoursDaysEditor({
 // ============================================================================
 // Rules Variant
 // ============================================================================
-
-interface HoursRule {
-  id: string;
-  days: number[];
-  start: string;
-  end: string;
-  description?: string;
-}
-
-function ruleSignature(slot: TimeSlot): string {
-  return `${slot.start}\u0000${slot.end}\u0000${slot.description ?? ''}`;
-}
-
-/** Collapse a schedule into rules: days sharing an identical time range group into one row. */
-export function scheduleToRules(schedule: DaySchedule[]): HoursRule[] {
-  const rulesBySignature = new Map<string, HoursRule[]>();
-  const orderedRules: HoursRule[] = [];
-
-  for (const day of schedule) {
-    for (const slot of day.hours) {
-      const signature = ruleSignature(slot);
-      const candidates = rulesBySignature.get(signature) ?? [];
-      // Duplicate identical slots on the same day each get their own rule so
-      // the schedule round-trips without dropping slots.
-      let rule = candidates.find((r) => !r.days.includes(day.day));
-      if (!rule) {
-        rule = {
-          id: generateId(),
-          days: [],
-          start: slot.start,
-          end: slot.end,
-          description: slot.description,
-        };
-        candidates.push(rule);
-        rulesBySignature.set(signature, candidates);
-        orderedRules.push(rule);
-      }
-      rule.days.push(day.day);
-    }
-  }
-
-  return orderedRules;
-}
-
-/** Expand rules into the canonical DaySchedule[] shape (all 7 days present). */
-export function rulesToSchedule(rules: HoursRule[]): DaySchedule[] {
-  return Array.from({ length: 7 }, (_, day) => ({
-    day,
-    hours: rules
-      .filter((rule) => rule.days.includes(day))
-      .map((rule) => ({
-        id: generateId(),
-        start: rule.start,
-        end: rule.end,
-        ...(rule.description ? { description: rule.description } : {}),
-      })),
-  }));
-}
 
 /** Normalized, id-free representation used to detect external value changes. */
 function normalizeSchedule(schedule: DaySchedule[]): string {
